requirements.txt
flask>=2.0.1
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
xgboost>=1.5.0
openai>=0.27.0
python-dotenv>=0.19.0
joblib>=1.0.1
pyyaml>=6.0 